Move *File to cfa/
