+++ src/lib/y2network/sysconfig/interfaces_writer.rb
@@ -20,8 +20,8 @@
 require "yast"
 require "y2network/udev_rule"
 require "yast2/execute"
-require "y2network/sysconfig/interface_file"
-require "y2network/sysconfig/routes_file"
+require "cfa/interface_file"
+require "cfa/routes_file"
 
 module Y2Network
   module Sysconfig
