+++ test/y2network/sysconfig/config_reader_test.rb
@@ -30,7 +30,7 @@ describe Y2Network::Sysconfig::ConfigReader do
   let(:connections) { [eth0_config] }
   let(:s390_devices) { [] }
   let(:drivers) { Y2Network::Driver.new("virtio_net", "") }
-  let(:routes_file) { instance_double(Y2Network::Sysconfig::RoutesFile, load: nil, routes: []) }
+  let(:routes_file) { instance_double(CFA::RoutesFile, load: nil, routes: []) }
   let(:dns_reader) { instance_double(Y2Network::Sysconfig::DNSReader, config: dns) }
   let(:hostname_reader) { instance_double(Y2Network::Sysconfig::HostnameReader, config: hostname) }
   let(:interfaces_reader) do
