+++ test/y2network/sysconfig/connection_config_readers/ctc_test.rb
@@ -19,7 +19,7 @@
 
 require_relative "../../../test_helper"
 require "y2network/sysconfig/connection_config_readers/ctc"
-require "y2network/sysconfig/interface_file"
+require "cfa/interface_file"
 require "y2network/boot_protocol"
 
 describe Y2Network::Sysconfig::ConnectionConfigReaders::Ctc do
@@ -34,7 +34,7 @@ describe Y2Network::Sysconfig::ConnectionConfigReaders::Ctc do
   let(:interface_name) { "ctc0" }
 
   let(:file) do
-    Y2Network::Sysconfig::InterfaceFile.find(interface_name).tap(&:load)
+    CFA::InterfaceFile.find(interface_name).tap(&:load)
   end
 
   describe "#connection_config" do
