+++ test/y2network/sysconfig/connection_config_readers/tap_test.rb
@@ -19,7 +19,7 @@
 
 require_relative "../../../test_helper"
 require "y2network/sysconfig/connection_config_readers/tap"
-require "y2network/sysconfig/interface_file"
+require "cfa/interface_file"
 require "y2network/interface_type"
 
 describe Y2Network::Sysconfig::ConnectionConfigReaders::Tap do
@@ -33,7 +33,7 @@ describe Y2Network::Sysconfig::ConnectionConfigReaders::Tap do
 
   let(:interface_name) { "tap0" }
   let(:file) do
-    Y2Network::Sysconfig::InterfaceFile.find(interface_name).tap(&:load)
+    CFA::InterfaceFile.find(interface_name).tap(&:load)
   end
 
   describe "#connection_config" do
