+++ test/y2network/sysconfig/connection_config_writers/dummy_test.rb
@@ -54,7 +54,7 @@ describe Y2Network::Sysconfig::ConnectionConfigWriters::Dummy do
     end
   end
 
-  let(:file) { Y2Network::Sysconfig::InterfaceFile.new(conn.name) }
+  let(:file) { CFA::InterfaceFile.new(conn.name) }
 
   describe "#write" do
     it "writes common properties" do
