+++ test/y2network/sysconfig/connection_config_writers/hsi_test.rb
@@ -36,7 +36,7 @@ describe Y2Network::Sysconfig::ConnectionConfigWriters::Hsi do
     end
   end
 
-  let(:file) { Y2Network::Sysconfig::InterfaceFile.new(conn.name) }
+  let(:file) { CFA::InterfaceFile.new(conn.name) }
 
   describe "#write" do
     it "writes common properties" do
