+++ test/y2network/sysconfig/connection_config_writers/tap_test.rb
@@ -56,7 +56,7 @@ describe Y2Network::Sysconfig::ConnectionConfigWriters::Tap do
     end
   end
 
-  let(:file) { Y2Network::Sysconfig::InterfaceFile.new(conn.name) }
+  let(:file) { CFA::InterfaceFile.new(conn.name) }
 
   describe "#write" do
     it "writes common properties" do
