+++ test/y2network/sysconfig/connection_config_writers/vlan_test.rb
@@ -40,7 +40,7 @@ describe Y2Network::Sysconfig::ConnectionConfigWriters::Vlan do
     end
   end
 
-  let(:file) { Y2Network::Sysconfig::InterfaceFile.new(conn.name) }
+  let(:file) { CFA::InterfaceFile.new(conn.name) }
 
   describe "#write" do
     it "writes the 'etherdevice' and the 'vlan_id' attributes" do
